改了一下数据分析的表结构和前端展示

+++ springboot-backend/src/main/java/com/datastructure/smartcanteen/entity/DishSalesAnalysis.java
@@ -19,7 +19,7 @@ public class DishSalesAnalysis {
     private LocalDate analysisDate;
     private Long dishId;
     private String dishName;
-    private Integer salesCount;
+    private BigDecimal salesCount;
     private BigDecimal salesAmount;
     private LocalDateTime createTime;
 }
